word/pyproject.toml
[project]
name = "word-polish"
version = "0.1.0"
description = "Word Polish MCP Server - 专业文字润色服务"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.0.8",
]

[project.scripts]
word-mcp = "word.server:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/word"]
